feat(concurrent): Refactor instrumentations per feature
